Change `npm` library build test to be on its own
