+++ deno.json
@@ -4,7 +4,7 @@
     "npm": "deno run -A ./scripts/npm/mod.js",
     "server": "deno run --allow-read --allow-net example/testServer.js",
     "buildServe": "deno task build && deno task server",
-    "test": "rm -rf D_state.sqlite && rm -rf __sysdb__.sqlite && deno task build && deno test --allow-read --allow-net --allow-write --trace-ops",
+    "test": "rm -rf D_state.sqlite && rm -rf __sysdb__.sqlite && deno task build && deno test --allow-read --allow-net --allow-write --allow-run --trace-ops",
     "jsdoc": "jsdoc src/history.js src/balance.js src/db.js src/mod.js"
   }
 }
